backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
requests
pymongo
